+++ apps/ledger-live-desktop/CHANGELOG.md
@@ -1,5 +1,25 @@
 # ledger-live-desktop
 
+## 2.49.2
+
+### Patch Changes
+
+-
+
+## 2.49.1
+
+### Patch Changes
+
+- Updated dependencies [[`455e43b34c`](https://github.com/LedgerHQ/ledger-live/commit/455e43b34c13ca7ed1d2920a36653caa250e42ab)]:
+  - @ledgerhq/live-common@27.5.0
+
+## 2.49.1-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies [[`455e43b34c`](https://github.com/LedgerHQ/ledger-live/commit/455e43b34c13ca7ed1d2920a36653caa250e42ab)]:
+  - @ledgerhq/live-common@27.5.0-hotfix.0
+
 ## 2.49.0
 
 ### Minor Changes